Try to recreate Tray Icon on Linux
After the org.kde.StatusNotifierWatcher disappears from the session
DBus (for example, when the GNOME Shell screen locker is engaged, or the
standalone system tray application is restarted), the tray icon must be
recreated when a new org.kde.StatusNotifierWatcher becomes available so
that the tray icon remains visible.

This commit achieves that by listening on the session DBus for name
ownership changes, and recreating the tray icon a after a short timeout
when the owner the the org.kde.StatusNotifierWatcher name changes. There
is unfortunately no way to see when is the new watcher ready to receive
new tray icons, so the timeout is hardcoded.

In the future, methods alternative to the built-in Electron API for Tray
icons might be investigated. Qt and GTK applications have no problem
maintaining their tray icons across watcher restarts, so there is
probably a more robust way to create tray icons than the one currently
offered by Electron.