Make @meetfranz/theme follow the accent color
We inject the value of the accent color into the themes, so the buttons
show the appropriate color. This allows removing some existing css
hacks.

In order to do this, the themes had to be turned into functions that
take the brand color (accent color) as an argument instead of
hard-coding it.

The css-based accent color for legacy components is left unchanged.