Restore ELECTRON_IS_DEV environment variable
As part of migrating to @electron/remote from electron.remote, 296ce5ce6
removed the electron-is-dev package and with it the support of selecting
dev/production mode with the ELECTRON_IS_DEV environmental variable.

This commit restores support for this variable. Because even the newest
version of the electron-is-dev package breaks in renderer processes, we
instead query the environment ourselves.