Refactor custom user agent handling logic
0. Any request header overrides in the recipe will still override the
   user agent string.
1. If the user has set a user agent override for the service, use that.
2. If the recipe has a predefined user agent, use that.
3. If the user has set a global user agent override, use that.
4. Otherwise, use the user agent string corresponding to the underlying
   Electron version.

If the current webpage is the Google login form, the we still remove the
Chrome version number from the user agent string.

The value that would be used if no custom override is
set (recipe-provided value or global override for the service, current
Electron instance for the global override) is shown as a placeholder in
the input box on the corresponding setting screen.