Move user agent override input boxes
These settings should hopefully be used only rarely, so they don't need
such a prominent place.

The global override is in Advanced, while the service override was moved
to the bottom of the service editor form (and now occupies the full
width of the form).

Show 'restart reqd' prompt for user-agent setting.